# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RT Finance Backend API",
        "description": "API Documentation for RT Finance Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api",
            "description": "Local Development"
        },
        {
            "url": "https://rt-finance-backend.vercel.app/api",
            "description": "Production"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/deferred-subscription": {
            "get": {
                "tags": [
                    "Deferred Subscription"
                ],
                "summary": "List all deferred subscriptions",
                "responses": {
                    "200": {
                        "description": "Array of subscriptions"
                    }
                }
            },
            "post": {
                "tags": [
                    "Deferred Subscription"
                ],
                "summary": "Create deferred subscription (prepaid)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "block": {
                                        "type": "string"
                                    },
                                    "houseNumber": {
                                        "type": "string"
                                    },
                                    "totalAmount": {
                                        "type": "integer"
                                    },
                                    "monthlyAmount": {
                                        "type": "integer"
                                    },
                                    "startMonth": {
                                        "type": "string",
                                        "example": "2026-01"
                                    },
                                    "endMonth": {
                                        "type": "string",
                                        "example": "2026-12"
                                    }
                                },
                                "required": [
                                    "block",
                                    "houseNumber",
                                    "totalAmount",
                                    "monthlyAmount",
                                    "startMonth",
                                    "endMonth"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription created"
                    }
                }
            }
        },
...
        "/deferred-subscription/{id}/deactivate": {
            "patch": {
                "tags": [
                    "Deferred Subscription"
                ],
                "summary": "Deactivate subscription",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription deactivated"
                    }
                }
            }
        },
        "/cron/run-ocr": {
            "post": {
                "tags": [
                    "Cron"
                ],
                "summary": "Run OCR batch processing",
                "description": "Requires x-cron-secret header",
                "parameters": [
                    {
                        "name": "x-cron-secret",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OCR processed"
                    }
                }
            }
        },
        "/cron/release-deferred-v1/{date}": {
            "post": {
                "tags": [
                    "Cron"
                ],
                "summary": "Release deferred subscriptions for month",
                "description": "Requires x-cron-secret header",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2026-01"
                    },
                    {
                        "name": "x-cron-secret",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deferred released"
                    }
                }
            }
        }
    }
}